import and show leads in admin.leads

+++ database/migrations/2020_07_03_193550_create_leads_table.php
@@ -15,7 +15,7 @@ class CreateLeadsTable extends Migration
     {
         Schema::create('leads', function (Blueprint $table) {
             $table->id();
-            $table->integer('document');
+            $table->bigInteger('document');
             $table->string('email')->nullable();
             $table->string('alternative_email')->nullable();
             $table->string('phone1',15)->nullable();
